% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grigspal.R
\name{get_palette}
\alias{get_palette}
\title{Get a color palette}
\usage{
get_palette(palette_name)
}
\arguments{
\item{palette_name}{A character string specifying the name of the palette.}
}
\value{
A vector of colors corresponding to the specified palette.
}
\description{
Retrieves a color palette by name.
}
