% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grigspal.R
\name{palette_peek}
\alias{palette_peek}
\title{Peek at a Color Palette}
\usage{
palette_peek(num_colors)
}
\arguments{
\item{num_colors}{An integer specifying the number of colors to display from the palette.}

\item{palette_name}{A character string specifying the name of the palette to preview.}
}
\value{
A plot showing the specified number of colors from the palette.
}
\description{
Displays a preview of a color palette with a specified number of colors.
}
